pyproject.toml
[tool.poetry]
name = "projeto-reconhecimento-facial"
version = "0.1.0"
description = ""
authors = ["Cecilius Sousa <ceciliusousa@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
cmake = "^3.31.1"
dlib = "^19.24.6"
face-recognition = "^1.3.0"
opencv-python = "^4.10.0.84"
mysql-connector-python = "^9.1.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
